ai-service/pyproject.toml
name = "academic-platform-ai"
version = "1.0.0"
description = "AI-powered semantic search service for academic platform"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
torch = "^2.1.0"
transformers = "^4.34.0"
sentence-transformers = "^2.2.0"
pymongo = "^4.6.0"
pydantic = "^2.4.0"
python-dotenv = "^1.0.0"
numpy = "^1.24.0"
scipy = "^1.11.0"
httpx = "^0.25.0"
python-multipart = "^0.0.6"
pydantic-settings = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
